Revert "Build homepage"
