+++ app/components/Footer/components/SocialsComponent.tsx
