+++ app/theme/elements/theme.ts
@@ -1,19 +1,19 @@
-import { DefaultTheme } from 'styled-components';
+import { DefaultTheme } from 'styled-components'
 
 export const theme: DefaultTheme = {
-  colors: {
-    primary: '20, 40, 59',
-    secondary: '3, 63 , 99',
-    hoverBackground: '3, 63, 99',
-    button: '204, 144, 12',
-    buttonText: '0, 0, 0',
-    buttonHover: '178, 116, 4',
-    buttonHoverText: '0, 0, 0',
-    text: '255, 255, 255',
-    secondaryText: '255, 255, 255',
-  },
-  windowWidth: 1320,
-  breakpoints: { desktop: 1320, tablet: 768, mobile: 375 },
-};
+    colors: {
+        primary: '0, 112, 243',
+        secondary: '28, 30 , 30',
+        hoverBackground: '255,255,255',
+        button: '0, 123, 255',
+        buttonText: '0,0,0',
+        buttonHover: '0, 86, 179',
+        buttonHoverText: '255, 255, 255',
+        text: '0,0,0',
+        secondaryText: '255,255,255',
+    },
+    windowWidth: 1320,
+    breakpoints: { desktop: 1320, tablet: 768, mobile: 375 },
+}
 
-export default theme;
+export default theme
