+++ app/theme/index.ts
@@ -1,2 +1,2 @@
-export { default as ThemeRegistry } from './elements/theme-registry';
-export { default as theme } from './elements/theme';
+export { default as ThemeRegistry } from './elements/ThemeRegistry'
+export { default as theme } from './elements/theme'
