+++ next.config.ts
@@ -1,10 +1,7 @@
-import type { NextConfig } from 'next';
+import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
-  compiler: {
-    styledComponents: true,
-  },
 };
 
 export default nextConfig;
