+++ package.json
@@ -12,7 +12,6 @@
     "next": "15.3.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-social-icons": "^6.24.0",
     "styled-components": "^6.1.18"
   },
   "devDependencies": {
